geocoding_service/requirements.txt
fastapi #библиотека для создания веб-приложений
requests
fastapi-cache2
